Reduce serialization size for version and variants

+++ synthesizer/src/block/transition/bytes.rs
@@ -34,7 +34,7 @@ impl<N: Network> FromBytes for Transition<N> {
         let function_name = FromBytes::read_le(&mut reader)?;
 
         // Read the number of inputs.
-        let num_inputs: u16 = FromBytes::read_le(&mut reader)?;
+        let num_inputs: u8 = FromBytes::read_le(&mut reader)?;
         // Read the inputs.
         let mut inputs = Vec::with_capacity(num_inputs as usize);
         for _ in 0..num_inputs {
@@ -43,7 +43,7 @@ impl<N: Network> FromBytes for Transition<N> {
         }
 
         // Read the number of outputs.
-        let num_outputs: u16 = FromBytes::read_le(&mut reader)?;
+        let num_outputs: u8 = FromBytes::read_le(&mut reader)?;
         // Read the outputs.
         let mut outputs = Vec::with_capacity(num_outputs as usize);
         for _ in 0..num_outputs {
@@ -58,7 +58,7 @@ impl<N: Network> FromBytes for Transition<N> {
             0 => None,
             1 => {
                 // Read the number of inputs for finalize.
-                let num_finalize_inputs = u16::read_le(&mut reader)?;
+                let num_finalize_inputs = u8::read_le(&mut reader)?;
                 // Read the inputs for finalize.
                 let mut finalize = Vec::with_capacity(num_finalize_inputs as usize);
                 for _ in 0..num_finalize_inputs {
@@ -103,12 +103,12 @@ impl<N: Network> ToBytes for Transition<N> {
         self.function_name.write_le(&mut writer)?;
 
         // Write the number of inputs.
-        (self.inputs.len() as u16).write_le(&mut writer)?;
+        (self.inputs.len() as u8).write_le(&mut writer)?;
         // Write the inputs.
         self.inputs.write_le(&mut writer)?;
 
         // Write the number of outputs.
-        (self.outputs.len() as u16).write_le(&mut writer)?;
+        (self.outputs.len() as u8).write_le(&mut writer)?;
         // Write the outputs.
         self.outputs.write_le(&mut writer)?;
 
@@ -122,7 +122,7 @@ impl<N: Network> ToBytes for Transition<N> {
                 // Write the finalize variant.
                 1u8.write_le(&mut writer)?;
                 // Write the number of inputs to finalize.
-                (finalize.len() as u16).write_le(&mut writer)?;
+                (finalize.len() as u8).write_le(&mut writer)?;
                 // Write the inputs to finalize.
                 finalize.write_le(&mut writer)?;
             }
